# planly-java/src/main/resources/static/openapi.yaml
openapi: 3.1.0
info:
  title: Planly Contracts PH0 T1
  version: 0.1.0
  description: >-
    Stack-A-Monolith: single Worker + Static Assets. API served via Hono under
    /v1, web served from web/dist via /* fallback. B2-01: generate, plans,
    readjust, operations active with 202 {opId,status} + GET /operations/:opId
    + GET /plans/:id teaser-or-full + POST /plans/:id/readjust 201
    {readjustmentId,baseVersion,opId}. CUT (probe-gated): entitlement, Steps
    2-6, discard. Cookie draftId: 30d, Secure, HttpOnly,
    SameSite Lax, Path /. Allowlist: 9 ops, 7 active, 2 x-probe-gated.
    Methods allowlist: GET, POST, PUT, DELETE only. Naming: id, opId,
    Idempotency-Key, readjustmentId, version, n. Teaser allowlist frozen.
    Idempotency-Key UUIDv4 only else 400.
  x-stack: Stack-A-Monolith single Worker + Static Assets, Hono /v1 + web/dist /*
  x-scope:
    go: PH0 T1 drafts + steps/:n Step1 Q1-Q4 plus B2-01 generate/plans/readjust/operations
    cut:
      - entitlement
      - steps-2-to-6
      - discard
  x-cookie-draftId:
    name: draftId
    maxAgeDays: 30
    secure: true
    httpOnly: true
    sameSite: Lax
    path: /
  x-slos:
    - id: SLO-1
      inv: INV-1
      text: Draft create p95 < 500ms, availability 99.9%
    - id: SLO-2
      inv: INV-2
      text: Step1 read p95 < 300ms, availability 99.9%
    - id: SLO-3
      inv: INV-3
      text: Step1 write p95 < 500ms, durability 99.999999%
    - id: SLO-4
      inv: INV-3
...
    - id: SLO-5
      inv: INV-4
...
servers:
...
paths:
...
  /drafts/{id}/steps/{n}:
    get:
      operationId: getDraftStep
      summary: Read draft step (active Step1 only, n=1)
      tags: [drafts]
      parameters:
        - $ref: '#/components/parameters/DraftIdPath'
        - $ref: '#/components/parameters/StepNPath'
      responses:
        '200':
          description: Step1 payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Step1Envelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '451':
          description: Steps 2-6 are probe-gated in PH0 T1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    put:
      operationId: putDraftStep
      summary: Write draft step (active Step1 only, n=1)
      tags: [drafts]
      parameters:
        - $ref: '#/components/parameters/DraftIdPath'
        - $ref: '#/components/parameters/StepNPath'
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Step1WriteRequest'
      responses:
        '200':
          description: Step1 saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Step1Envelope'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: version conflict
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '451':
          description: Steps 2-6 are probe-gated in PH0 T1
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /plans/{id}:
    get:
      operationId: getPlan
      summary: Read plan teaser-or-full with ETag
      tags: [gated]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Id'
        - $ref: '#/components/parameters/IfNoneMatch'
      responses:
        '200':
          description: Plan payload; unentitled callers get full:null teaser path
          headers:
            ETag:
              schema:
                type: string
                example: plan-dr_01HXYZ-v1
              description: ETag plan-<id>-vN
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PlanResponse'
        '304':
          description: Not modified (If-None-Match hit)
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden (bound-draft access only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          $ref: '#/components/responses/NotFound'
  /plans/{id}/readjust:
    post:
      operationId: readjustPlan
      summary: Append plan readjustment (baseVersion required)
      tags: [gated]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            $ref: '#/components/schemas/Id'
        - $ref: '#/components/parameters/IdempotencyKeyRequired'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReadjustRequest'
      responses:
        '201':
          description: Readjustment appended; poll opId for new plan version
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReadjustAccepted'
        '400':
          description: Missing baseVersion or malformed Idempotency-Key (UUIDv4 only)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthenticated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '402':
          description: Entitlement required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Unknown or stale baseVersion
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Delta invalid
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
components:
  parameters:
    DraftIdPath:
      name: id
      in: path
      required: true
      description: Draft id (naming canonical)
      schema:
        $ref: '#/components/schemas/Id'
    StepNPath:
      name: n
      in: path
      required: true
      description: Step number 1..6; only n=1 active in PH0 T1, 2-6 gated
      schema:
        type: integer
        minimum: 1
        maximum: 6
        example: 1
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: false
      description: Idempotency key, UUIDv4 only else 400
      schema:
        type: string
        format: uuid
        pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$
        example: 550e8400-e29b-41d4-a716-446655440000
    IdempotencyKeyRequired:
      name: Idempotency-Key
      in: header
      required: true
      description: Idempotency key required on generate/readjust, UUIDv4 only else 400
      schema:
        type: string
        format: uuid
        pattern: ^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-4[0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}$
        example: 550e8400-e29b-41d4-a716-446655440000
...
    IfNoneMatch:
      name: If-None-Match
      in: header
      required: false
      description: Conditional GET with plan ETag plan-<id>-vN yields 304
      schema:
        type: string
        example: plan-dr_01HXYZ-v1
...
  schemas:
    Id:
      type: string
      minLength: 1
      maxLength: 128
      example: dr_01HXYZ
    OpId:
      type: string
      description: Operation id canonical (opId)
      minLength: 1
      maxLength: 128
      example: op_01HXYZ
    Version:
      type: integer
      minimum: 1
      example: 1
    StepN:
      type: integer
      minimum: 1
      maximum: 6
      example: 1
    ReadjustmentId:
      type: string
      description: readjustmentId canonical
      minLength: 1
      maxLength: 128
      example: ra_01HXYZ
    Step1:
      type: object
      required: [q1, q2, q3, q4]
      additionalProperties: false
      properties:
        q1:
          type: string
          minLength: 1
          maxLength: 2000
          description: Q1 goal
          example: Launch a weekly study plan
        q2:
          type: string
          minLength: 1
          maxLength: 2000
          description: Q2 context
          example: Evenings only, 30 min per day
        q3:
          type: string
          minLength: 1
          maxLength: 2000
          description: Q3 constraints
          example: No weekends, low cost
        q4:
          type: string
          minLength: 1
          maxLength: 2000
          description: Q4 success criteria
          example: Finish two chapters per week
    Step1WriteRequest:
      type: object
      required: [version, step1]
      additionalProperties: false
      properties:
        version:
          $ref: '#/components/schemas/Version'
        opId:
          $ref: '#/components/schemas/OpId'
        step1:
          $ref: '#/components/schemas/Step1'
    Step1Envelope:
      type: object
      required: [id, version, n, step1]
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        version:
          $ref: '#/components/schemas/Version'
        n:
          $ref: '#/components/schemas/StepN'
        opId:
          $ref: '#/components/schemas/OpId'
        step1:
          $ref: '#/components/schemas/Step1'
...
    ReadjustRequest:
      type: object
      required: [baseVersion, delta]
      additionalProperties: false
      properties:
        baseVersion:
          $ref: '#/components/schemas/Version'
        delta:
          $ref: '#/components/schemas/ReadjustDelta'
        opId:
          $ref: '#/components/schemas/OpId'
    ReadjustDelta:
      type: object
      required: [missedDate]
      additionalProperties: false
      description: confidence outside 1-5 -> 400 with field-level error
      properties:
        missedDate:
          type: string
          format: date
          description: Missed calendar date YYYY-MM-DD
          example: 2026-09-20
        taskId:
          $ref: '#/components/schemas/Id'
        confidence:
          type: integer
          minimum: 1
          maximum: 5
        evidence:
          type: string
    ReadjustAccepted:
      type: object
      required: [readjustmentId, baseVersion, opId]
      additionalProperties: false
      properties:
        readjustmentId:
          $ref: '#/components/schemas/ReadjustmentId'
        baseVersion:
          $ref: '#/components/schemas/Version'
        opId:
          $ref: '#/components/schemas/OpId'
...
    PlanTeaser:
      type: object
      additionalProperties: false
      description: Teaser allowlist frozen; unlisted keys default to full
      properties:
        title:
          type: string
          maxLength: 120
        summary:
          type: string
          maxLength: 280
        stepCount:
          type: integer
          minimum: 1
        priceRange:
          type: string
    Sprint:
      type: object
      required: [id, n, title]
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        n:
          type: integer
          minimum: 1
        title:
          type: string
          maxLength: 120
        dailyIds:
          type: array
          items:
            $ref: '#/components/schemas/Id'
    Daily:
      type: object
      required: [id, date, sprintId]
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        date:
          type: string
          format: date
        sprintId:
          $ref: '#/components/schemas/Id'
        tasks:
          type: array
          description: Missing taskId/title/done_criteria -> 400 with field-level error
          items:
            $ref: '#/components/schemas/DailyTask'
    DailyTask:
      type: object
      required: [taskId, title, done_criteria]
      additionalProperties: false
      description: Missing taskId/title/done_criteria -> 400 with field-level error
      properties:
        taskId:
          $ref: '#/components/schemas/Id'
        doing_verb:
          type: string
        title:
          type: string
        minutes:
          type: integer
          minimum: 1
        done_criteria:
          type: string
        planly_check:
          type: string
        confidence:
          type: integer
          minimum: 1
          maximum: 5
        miss_rule:
          type: string
...
    Track:
      type: object
      required: [id, title, level, sprint, taskIds]
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        title:
          type: string
        level:
          type: string
          enum: [L0, L1, L2, L3, L4]
        sprint:
          $ref: '#/components/schemas/Id'
        taskIds:
          type: array
          items:
            $ref: '#/components/schemas/Id'
    PlanFullMetrics:
      type: object
      required: [totalSprints, subjects, studyHours]
      additionalProperties: false
      properties:
        totalSprints:
          type: integer
          minimum: 1
        subjects:
          type: integer
          minimum: 1
        studyHours:
          type: number
          minimum: 0
        estDuration:
          type: string
        estCompletion:
          type: string
          format: date
    PlanFull:
      type: object
      required: [sprints, dailies, metrics, tracks, progress]
      additionalProperties: false
      properties:
        sprints:
          type: array
          items:
            $ref: '#/components/schemas/Sprint'
        dailies:
          type: array
          items:
            $ref: '#/components/schemas/Daily'
        metrics:
          $ref: '#/components/schemas/PlanFullMetrics'
        tracks:
          type: array
          items:
            $ref: '#/components/schemas/Track'
        progress:
          $ref: '#/components/schemas/PlanProgress'
    PlanProgress:
      type: object
      required: [done, total]
      additionalProperties: false
      properties:
        done:
          type: integer
          minimum: 0
        total:
          type: integer
          minimum: 0
    PlanResponse:
      type: object
      required: [id, version, teaser, full]
      additionalProperties: false
      properties:
        id:
          $ref: '#/components/schemas/Id'
        version:
          $ref: '#/components/schemas/Version'
        teaser:
          $ref: '#/components/schemas/PlanTeaser'
        full:
          anyOf:
            - $ref: '#/components/schemas/PlanFull'
            - type: 'null'
          description: full:null explicit on teaser path
    Error:
      type: object
      required: [code, message]
      additionalProperties: false
      properties:
        code:
          type: string
          example: gated
        message:
          type: string
          example: Probe-gated in PH0 T1
    # T1 wizard Steps 2-6 StepDef skeletons — static-string grade,
    # rendered-unverified. SSR Q1-Q4 (Step1) untouched. CUT LINE: no horizon
    # field (2-12mo is subscription-validity copy only), no sprint/daily
    # object shapes (loader/metric strings are display-only).
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'